Conexión mongodb

+++ src/cors.rs
@@ -1,13 +1,10 @@
 use rocket::http::Method;
-use rocket_cors::AllowedHeaders;
-
 use rocket_cors;
-
+use rocket_cors::AllowedHeaders;
 use rocket_cors::AllowedOrigins;
-
 use rocket_cors::Cors;
 
-pub fn run(port: &i32) -> Cors {
+pub fn run(port: &u16) -> Cors {
     let allowed_origins = AllowedOrigins::some_exact(&[
         format!("http://localhost:{port}"),
         format!("http://127.0.0.1:{port}"),
